Reparied transport, however the typescript object is not being instantiated. There is only a null object reference named thisInspection being passed to
transpot.SaveInspection(thisInspection: NewInspection);

I have verified the transport of the object will work once I get the TS syntax down for creating an object in this way. If I create the constructor to take 3 parameters, and create an object and pass the three to it, I get an "object is not a constructor" error.

Object is being created in app.ts in SaveInspection onclick handler.