Updates to multiple files to prevent the scheduling calendar from being built completely if the result for a final inspection for that permit is an 'A' or 'P', then informing the user of a permit scheduling issue. Checks in the server code prevent changes made in the client scripts from allowing false data to be used to schedule inspections.
Functionality added to check the master permit to determine whether a private provider inspector is being used (a bit on the master permit table based on permit type), and insert a row into the private providers table if the test is true; and inserting a row into the inspection table along with data for an additional field.

Updates to the NewInspection.Cancel functions SQL. The result of both inspection and private provider inspections is set to 'C' in order to facilitate tracking.

Removed dead code and unnecessary comments, tested internal and external functionality.

Added addition html, css, and js code giving the user the ability to hide the issues list displayed after a failed attempt to save an inspection. Consideration for the possibilty of adding a timer to automatically hide the issues list after a few seconds is ongoing.

Updated function names to reflect changes in functionality, (eg. Delete() changed to Cancel() ).

New updates to IsExternalUser boolean checks. Using new functionality to check for internal id in order to determine if the user is on the internet or intranet.