update UI to handle Permits if they search for a void permit (permit exists in DB but is not included in Permit List)
Update Hold.cs to exclude all Holds not SatFinal or not SatInspection.

Update Permit To add error text to master if Assoc permit has a hold or charge.

Updated Permit Controller to prevent URL link on project address from being added as a link if IsExternalUser is true, even if IsSupervisor is true.